requirements.txt
streamlit 
python-dotenv
pyqrcode
pyzbar
opencv-python
pypng
azure-cosmos
reportlab